requirements.txt
google-api-python-client>=1.9.3
oauth2client
